+++ packages/sdk/CHANGELOG.md
@@ -13,6 +13,8 @@ commit that affects the code in a way that consumers might care about should inc
 
 -   Add workspace id to base object.
 -   Add `Field.updateNameAsync`
+-   Fixes a crash when a linked record is deleted
+-   Fixes a crash when a record containing a linked record is deleted and restored
 
 ## [1.10.2](https://github.com/airtable/blocks/compare/@airtable/blocks@1.10.1...@airtable/blocks@1.10.2) - 2022-01-25